functions/requirements.txt
flask
functions-framework
openai
google-cloud-firestore
gspread
oauth2client
google-cloud-secret-manager
